Refactoring, commenting
