+++ services/web/src/components/pages/ApplyV2/components/forms/Eligibility/components/VerifyAddress/index.tsx
@@ -1,9 +1,10 @@
-import { Alert, Card, Button, Form, Select, Input } from 'antd';
+import { Card, Button, Form, Select, Input } from 'antd';
 
-import { useSelector } from 'react-redux';
+// Utils
+import parseAddress from './utils/parseAddress';
+import formatAddress from './utils/formatAddress';
 
-import { useState } from 'react';
-
+// Custom hook used to interact with the SmartyStreets autosuggestions API
 import useAddressAutoSuggestions from '../../../../../../../../utils/hooks/useAddressAutoSuggestions';
 
 const { Option } = Select;
@@ -12,35 +13,9 @@ export default function Index({
   formValues,
   setFormValues,
   handleChange,
-  onStateChange,
   setEligibilityContent,
-  currentUser,
-  dispatch,
 }) {
-  const request = useSelector(state => state.requests.request);
-
-  const [isAlertOpen, setIsAlertOpen] = useState(false);
-
-  const [autosuggestions, handleAutoChange] = useAddressAutoSuggestions([]);
-
-  const [address, setAddress] = useState({});
-
-  const parseAddress = str => {
-    str = str.split(',');
-
-    let address = str[0].trim();
-    let cityName = str[1].trim();
-    let state = str[2].trim();
-    let zipCode = str[3].trim();
-
-    return [address, cityName, state, zipCode];
-  };
-
-  const handleChangeWrapper = e => {
-    setIsAlertOpen(false);
-
-    return handleChange(e);
-  };
+  const [autosuggestions, handleSearch] = useAddressAutoSuggestions([]);
 
   const handleAddressChange = str => {
     let [address, cityName, state, zipCode] = parseAddress(str);
@@ -59,11 +34,7 @@ export default function Index({
   };
 
   return (
-    <Form
-      layout="vertical"
-      onChange={handleChangeWrapper}
-      onFinish={handleFinish}
-    >
+    <Form layout="vertical" onFinish={handleFinish}>
       <Card>
         <Form.Item
           label="Address"
@@ -76,7 +47,7 @@ export default function Index({
             size="large"
             placeholder="Address"
             onChange={handleAddressChange}
-            onSearch={handleAutoChange}
+            onSearch={handleSearch}
           >
             {autosuggestions.map(address => (
               <Option value={formatAddress(address)}>
@@ -92,14 +63,10 @@ export default function Index({
           label="Address Line Two"
           name="addressLine2"
         >
-          <Input name="addressLine2" />
+          <Input onChange={handleChange} name="addressLine2" />
         </Form.Item>
         <Button htmlType="submit">Next</Button>
       </Card>
     </Form>
   );
 }
-
-const formatAddress = address => {
-  return `${address.streetLine}, ${address.city}, ${address.state}, ${address.zipcode}`;
-};
