+++ services/web/src/components/pages/ApplyV2/components/forms/Eligibility/components/VerifyAddress/utils/formatAddress.ts
@@ -1,0 +1,5 @@
+const formatAddress = address => {
+  return `${address.streetLine}, ${address.city}, ${address.state}, ${address.zipcode}`;
+};
+
+export default formatAddress;
