+++ services/web/src/components/pages/ApplyV2/components/forms/Eligibility/components/VerifyAddress/utils/parseAddress.ts
@@ -1,0 +1,12 @@
+const parseAddress = str => {
+  str = str.split(',');
+
+  let address = str[0].trim();
+  let cityName = str[1].trim();
+  let state = str[2].trim();
+  let zipCode = str[3].trim();
+
+  return [address, cityName, state, zipCode];
+};
+
+export default parseAddress;
